# docs/openapi.yaml
openapi: 3.1.0
...
info:
  title: Artly API
  version: 1.0.0
  summary: REST API cho mạng xã hội bài thi vẽ Artly
  description: |
    Hợp đồng API cho MVP Artly, nơi học sinh và giáo viên đăng bài thi vẽ,
    thả reaction, bình luận, nhắn tin trực tiếp và trò chuyện với trợ lý AI đa
    năng.

    Các endpoint được bảo vệ chấp nhận access token qua
    `Authorization: Bearer <token>`. Khi chạy local/testing hoặc chế độ demo,
    client có thể dùng `X-User-ID` để chọn một tài khoản mẫu; header này **không
    phải xác thực dùng cho production**.

    Mọi thuộc tính JSON dùng camelCase. Mọi phản hồi lỗi đều có cùng cấu trúc:
    `{ "error": { "code", "message", "details" } }`.

    Mọi ID tài nguyên trong JSON, header, path và query là chuỗi UUID. Các giá
    trị phân trang, số lượng reaction/bình luận, kết quả đếm và mã trạng thái
    HTTP vẫn là số.
servers:
  - url: http://localhost:3000
    description: Goravel chạy trên máy local
...
paths:
...
  /api/v1/demo/sessions:
    post:
      tags:
        - Người dùng
      summary: Đăng nhập tài khoản demo
      description: |
        Xác thực username/password của tài khoản demo và trả về user dùng cho
        các request MVP tiếp theo qua header `X-User-ID`.
      operationId: createDemoSession
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DemoLoginRequest"
            examples:
              cafeDemo:
                summary: Đăng nhập học sinh demo
                value:
                  username: thu.ha.cafe
                  password: artly-demo
      responses:
        "201":
          description: Đăng nhập demo thành công.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
              examples:
                signedIn:
                  summary: User hiện tại
                  value:
                    data:
                      id: "00000000-0000-4000-8000-000000000001"
                      username: thu.ha.cafe
                      displayName: Lê Thu Hà
                      role: STUDENT
                      avatarUrl: null
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/InvalidDemoCredentials"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/users:
    get:
      tags:
        - Người dùng
      summary: Lấy danh sách tài khoản mẫu
      description: |
        Trả về tài khoản theo thứ tự tăng dần của UUID. Endpoint này không yêu
        cầu `X-User-ID` để người dùng có thể chọn tài khoản demo.
      operationId: listUsers
      security: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/DirectoryPageSize"
      responses:
        "200":
          description: Danh sách tài khoản mẫu và thông tin phân trang.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
              examples:
                firstPage:
                  summary: Học sinh và giáo viên mẫu
                  value:
                    data:
                      - id: "00000000-0000-4000-8000-000000000001"
                        username: minh.an
                        displayName: Trần Minh An
                        role: STUDENT
                        avatarUrl: https://i.pravatar.cc/256?img=12
                      - id: "00000000-0000-4000-8000-000000000002"
                        username: co.lan
                        displayName: Cô Nguyễn Hoài Lan
                        role: TEACHER
                        avatarUrl: https://i.pravatar.cc/256?img=47
                      - id: "00000000-0000-4000-8000-000000000003"
                        username: phuong.thao
                        displayName: Nguyễn Phương Thảo
                        role: STUDENT
                        avatarUrl: https://i.pravatar.cc/256?img=32
                    pagination:
                      page: 1
                      pageSize: 20
                      totalItems: 3
                      totalPages: 1
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/topics:
    get:
      tags:
        - Chủ đề
      summary: Lấy danh sách chủ đề
      description: |
        Trả về các chủ đề chuẩn hóa. Mảng `aliases` giúp giao diện hiển thị những
        cách gọi tương đương mà trợ lý có thể nhận diện.
      operationId: listTopics
      security: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/DirectoryPageSize"
      responses:
        "200":
          description: Danh sách chủ đề và thông tin phân trang.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopicListResponse"
              examples:
                drawingTopics:
                  summary: Một số chủ đề vẽ
                  value:
                    data:
                      - id: "10000000-0000-4000-8000-000000000001"
                        slug: phong-canh
                        name: Phong cảnh
                        aliases:
                          - Cảnh vật
                          - Landscape
                      - id: "10000000-0000-4000-8000-000000000002"
                        slug: chan-dung
                        name: Chân dung
                        aliases:
                          - Vẽ người
                          - Portrait
                      - id: "10000000-0000-4000-8000-000000000003"
                        slug: moi-truong
                        name: Môi trường
                        aliases:
                          - Bảo vệ môi trường
                          - Environment
                      - id: "10000000-0000-4000-8000-000000000004"
                        slug: hoa-binh
                        name: Hòa bình
                        aliases:
                          - Tình bạn
                          - Peace
                      - id: "10000000-0000-4000-8000-000000000005"
                        slug: di-san-van-hoa
                        name: Di sản văn hóa
                        aliases:
                          - Di sản
                          - Heritage
                      - id: "10000000-0000-4000-8000-000000000006"
                        slug: uoc-mo
                        name: Ước mơ
                        aliases:
                          - Giấc mơ
                          - Dream
                    pagination:
                      page: 1
                      pageSize: 20
                      totalItems: 6
                      totalPages: 1
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/posts:
    get:
      tags:
        - Bài viết
      summary: Lấy bảng tin bài thi vẽ
      description: |
        Trả về bài viết mới nhất trước. Khi có `topicId`, chỉ những bài được gắn
        chủ đề tương ứng mới được trả về. `viewerHasReacted` được tính theo tài
        khoản hiện tại.
      operationId: listPosts
      security:
        - DemoUserId: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PostPageSize"
        - $ref: "#/components/parameters/TopicId"
      responses:
        "200":
          description: Bảng tin và thông tin phân trang.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostListResponse"
              examples:
                landscapeFeed:
                  summary: Bảng tin chủ đề phong cảnh
                  value:
                    data:
                      - id: "20000000-0000-4000-8000-000000000005"
                        title: Thành phố trên mây
                        caption: Em tưởng tượng một thành phố xanh có những khu vườn bay giữa bầu trời.
                        imageUrl: https://images.unsplash.com/photo-1541961017774-22349e4a1262?auto=format&fit=crop&w=1200&q=80
                        examName: Ước mơ của em
                        author:
                          id: "00000000-0000-4000-8000-000000000001"
                          username: minh.an
                          displayName: Trần Minh An
                          role: STUDENT
                          avatarUrl: https://i.pravatar.cc/256?img=12
                        topics:
                          - id: "10000000-0000-4000-8000-000000000001"
                            slug: phong-canh
                            name: Phong cảnh
                            aliases:
                              - Cảnh vật
                              - Landscape
                          - id: "10000000-0000-4000-8000-000000000006"
                            slug: uoc-mo
                            name: Ước mơ
                            aliases:
                              - Giấc mơ
                              - Dream
                        reactionCount: 1
                        commentCount: 2
                        viewerHasReacted: false
                        createdAt: "2026-07-24T08:30:00+07:00"
                      - id: "20000000-0000-4000-8000-000000000001"
                        title: Mầm xanh tương lai
                        caption: Bài vẽ màu nước về những bàn tay cùng nâng niu một mầm cây xanh.
                        imageUrl: http://localhost:5173/demo-art/mam-xanh-tuong-lai.png
                        examName: Sắc màu xanh 2026
                        author:
                          id: "00000000-0000-4000-8000-000000000001"
                          username: minh.an
                          displayName: Trần Minh An
                          role: STUDENT
                          avatarUrl: https://i.pravatar.cc/256?img=12
                        topics:
                          - id: "10000000-0000-4000-8000-000000000001"
                            slug: phong-canh
                            name: Phong cảnh
                            aliases:
                              - Cảnh vật
                              - Landscape
                          - id: "10000000-0000-4000-8000-000000000003"
                            slug: moi-truong
                            name: Môi trường
                            aliases:
                              - Bảo vệ môi trường
                              - Environment
                        reactionCount: 2
                        commentCount: 1
                        viewerHasReacted: false
                        createdAt: "2026-07-24T08:30:00+07:00"
                    pagination:
                      page: 1
                      pageSize: 10
                      totalItems: 2
                      totalPages: 1
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/DemoIdentityRequired"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/RequestTimeout"

    post:
      tags:
        - Bài viết
      summary: Đăng một bài thi vẽ
      description: |
        Tạo bài viết cho tài khoản hiện tại. MVP chỉ nhận URL ảnh; endpoint này
        không nhận file upload.
      operationId: createPost
      security:
        - DemoUserId: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequest"
            examples:
              watercolorEntry:
                summary: Bài dự thi màu nước
                value:
                  title: Mầm xanh tương lai
                  caption: Bài vẽ màu nước về những bàn tay cùng nâng niu một mầm cây xanh.
                  imageUrl: http://localhost:5173/demo-art/mam-xanh-tuong-lai.png
                  examName: Sắc màu xanh 2026
                  topicIds:
                    - "10000000-0000-4000-8000-000000000001"
                    - "10000000-0000-4000-8000-000000000003"
      responses:
        "201":
          description: Đã tạo bài viết.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostResponse"
              examples:
                created:
                  summary: Bài viết vừa tạo
                  value:
                    data:
                      id: "20000000-0000-4000-8000-000000000001"
                      title: Mầm xanh tương lai
                      caption: Bài vẽ màu nước về những bàn tay cùng nâng niu một mầm cây xanh.
                      imageUrl: http://localhost:5173/demo-art/mam-xanh-tuong-lai.png
                      examName: Sắc màu xanh 2026
                      author:
                        id: "00000000-0000-4000-8000-000000000001"
                        username: minh.an
                        displayName: Trần Minh An
                        role: STUDENT
                        avatarUrl: https://i.pravatar.cc/256?img=12
                      topics:
                        - id: "10000000-0000-4000-8000-000000000001"
                          slug: phong-canh
                          name: Phong cảnh
                          aliases:
                            - Cảnh vật
                            - Landscape
                        - id: "10000000-0000-4000-8000-000000000003"
                          slug: moi-truong
                          name: Môi trường
                          aliases:
                            - Bảo vệ môi trường
                            - Environment
                      reactionCount: 0
                      commentCount: 0
                      viewerHasReacted: false
                      createdAt: "2026-07-24T08:30:00+07:00"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/DemoIdentityRequired"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/RequestTimeout"
  /api/v1/posts/{id}/comments:
    parameters:
      - $ref: "#/components/parameters/PostId"

    get:
      tags:
        - Bài viết
      summary: Lấy bình luận của bài viết
      description: |
        Trả về bình luận phẳng theo thứ tự mới nhất trước. Tác giả có thể xóa
        bình luận của chính mình; MVP chưa hỗ trợ trả lời, sửa hoặc reaction cho
        bình luận.
      operationId: listPostComments
      security:
        - BearerAuth: []
        - DemoUserId: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/CommentPageSize"
      responses:
        "200":
          description: Danh sách bình luận và thông tin phân trang.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentListResponse"
              examples:
                newestComments:
                  summary: Bình luận mới nhất của bài viết
                  value:
                    data:
                      - id: "30000000-0000-4000-8000-000000000001"
                        postId: "20000000-0000-4000-8000-000000000001"
                        body: Em rất thích cách bạn phối màu cho phần bầu trời.
                        author:
                          id: "00000000-0000-4000-8000-000000000002"
                          username: co.lan
                          displayName: Cô Nguyễn Hoài Lan
                          role: TEACHER
                          avatarUrl: https://i.pravatar.cc/256?img=47
                        createdAt: "2026-07-25T09:15:00+07:00"
                    pagination:
                      page: 1
                      pageSize: 20
                      totalItems: 1
                      totalPages: 1
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/DemoIdentityRequired"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags:
        - Bài viết
      summary: Bình luận vào bài viết
      description: |
        Tạo một bình luận phẳng của tài khoản hiện tại. Backend loại bỏ khoảng
        trắng Unicode ở đầu và cuối trước khi kiểm tra độ dài từ 1 đến 3000 ký
        tự Unicode.
      operationId: createPostComment
      security:
        - BearerAuth: []
        - DemoUserId: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentRequest"
            examples:
              teacherFeedback:
                summary: Nhận xét của giáo viên
                value:
                  body: Em rất thích cách bạn phối màu cho phần bầu trời.
      responses:
        "201":
          description: Đã tạo bình luận.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentResponse"
              examples:
                created:
                  summary: Bình luận vừa tạo
                  value:
                    data:
                      id: "30000000-0000-4000-8000-000000000001"
                      postId: "20000000-0000-4000-8000-000000000001"
                      body: Em rất thích cách bạn phối màu cho phần bầu trời.
                      author:
                        id: "00000000-0000-4000-8000-000000000002"
                        username: co.lan
                        displayName: Cô Nguyễn Hoài Lan
                        role: TEACHER
                        avatarUrl: https://i.pravatar.cc/256?img=47
                      createdAt: "2026-07-25T09:15:00+07:00"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/DemoIdentityRequired"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/posts/{id}/comments/{commentId}:
    parameters:
      - $ref: "#/components/parameters/PostId"
      - $ref: "#/components/parameters/CommentId"

    delete:
      tags:
        - Bài viết
      summary: Xóa bình luận của chính mình
      description: |
        Soft-delete một bình luận còn hiển thị khi bình luận thuộc bài viết
        trong path và tác giả là tài khoản hiện tại. API trả 404 nếu bình luận
        không tồn tại, đã bị xóa, không thuộc bài viết hoặc không thuộc người
        gọi để không làm lộ dữ liệu. `id` hoặc `commentId` không phải UUID hợp
        lệ sẽ trả 400 `BAD_REQUEST`.
      operationId: deletePostComment
      security:
        - BearerAuth: []
        - DemoUserId: []
      responses:
        "204":
          description: Đã xóa bình luận; response không có body.
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/DemoIdentityRequired"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/posts/{id}/reaction:
    parameters:
      - $ref: "#/components/parameters/PostId"

    put:
      tags:
        - Bài viết
      summary: Thả reaction cho bài viết
      description: |
        Thêm reaction của tài khoản hiện tại vào bài viết. Thao tác có tính
        idempotent: gọi lại nhiều lần vẫn chỉ tạo một reaction cho cặp bài
        viết-người dùng.
      operationId: putPostReaction
      security:
        - DemoUserId: []
      responses:
        "200":
          description: Trạng thái reaction hiện tại của bài viết.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReactionResponse"
              examples:
                reacted:
                  summary: Đã thả tim
                  value:
                    data:
                      reactionCount: 3
                      viewerHasReacted: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/DemoIdentityRequired"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"

    delete:
      tags:
        - Bài viết
      summary: Gỡ reaction khỏi bài viết
      description: |
        Gỡ reaction của tài khoản hiện tại khỏi bài viết. Thao tác có tính
        idempotent: nếu reaction không tồn tại, API vẫn trả trạng thái thành công.
      operationId: deletePostReaction
      security:
        - DemoUserId: []
      responses:
        "200":
          description: Trạng thái reaction hiện tại của bài viết.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReactionResponse"
              examples:
                removed:
                  summary: Đã gỡ tim
                  value:
                    data:
                      reactionCount: 2
                      viewerHasReacted: false
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/DemoIdentityRequired"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/messages:
    get:
      tags:
        - Tin nhắn
      summary: Lấy cuộc trò chuyện với một người dùng
      description: |
        Trả về tin nhắn giữa tài khoản hiện tại và `peerId`, mới nhất trước để
        hỗ trợ REST polling. API chỉ trả những tin nhắn mà tài khoản hiện tại là
        người gửi hoặc người nhận.
      operationId: listMessages
      security:
        - DemoUserId: []
      parameters:
        - $ref: "#/components/parameters/PeerId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/MessagePageSize"
      responses:
        "200":
          description: Các tin nhắn trong cuộc trò chuyện và thông tin phân trang.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageListResponse"
              examples:
                conversation:
                  summary: Cuộc trò chuyện giữa học sinh và giáo viên
                  value:
                    data:
                      - id: "50000000-0000-4000-8000-000000000002"
                        sender:
                          id: "00000000-0000-4000-8000-000000000002"
                          username: co.lan
                          displayName: Cô Nguyễn Hoài Lan
                          role: TEACHER
                          avatarUrl: https://i.pravatar.cc/256?img=47
                        receiver:
                          id: "00000000-0000-4000-8000-000000000001"
                          username: minh.an
                          displayName: Trần Minh An
                          role: STUDENT
                          avatarUrl: https://i.pravatar.cc/256?img=12
                        body: Em thử tăng độ tương phản ở vùng tiền cảnh nhé, bố cục đang rất tốt.
                        createdAt: "2026-07-24T09:10:00+07:00"
                      - id: "50000000-0000-4000-8000-000000000001"
                        sender:
                          id: "00000000-0000-4000-8000-000000000001"
                          username: minh.an
                          displayName: Trần Minh An
                          role: STUDENT
                          avatarUrl: https://i.pravatar.cc/256?img=12
                        receiver:
                          id: "00000000-0000-4000-8000-000000000002"
                          username: co.lan
                          displayName: Cô Nguyễn Hoài Lan
                          role: TEACHER
                          avatarUrl: https://i.pravatar.cc/256?img=47
                        body: Cô xem giúp em phần phối màu của bài Mầm xanh tương lai với ạ.
                        createdAt: "2026-07-24T09:10:00+07:00"
                    pagination:
                      page: 1
                      pageSize: 50
                      totalItems: 2
                      totalPages: 1
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/DemoIdentityRequired"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags:
        - Tin nhắn
      summary: Gửi tin nhắn trực tiếp
      description: |
        Gửi tin nhắn từ tài khoản hiện tại đến một tài khoản mẫu khác.
      operationId: createMessage
      security:
        - DemoUserId: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageRequest"
            examples:
              feedbackQuestion:
                summary: Học sinh hỏi nhận xét
                value:
                  recipientId: "00000000-0000-4000-8000-000000000002"
                  body: Cô xem giúp em phần phối màu của bài Mầm xanh tương lai với ạ.
      responses:
        "201":
          description: Đã gửi tin nhắn.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
              examples:
                sent:
                  summary: Tin nhắn vừa gửi
                  value:
                    data:
                      id: "50000000-0000-4000-8000-000000000001"
                      sender:
                        id: "00000000-0000-4000-8000-000000000001"
                        username: minh.an
                        displayName: Trần Minh An
                        role: STUDENT
                        avatarUrl: https://i.pravatar.cc/256?img=12
                      receiver:
                        id: "00000000-0000-4000-8000-000000000002"
                        username: co.lan
                        displayName: Cô Nguyễn Hoài Lan
                        role: TEACHER
                        avatarUrl: https://i.pravatar.cc/256?img=47
                      body: Cô xem giúp em phần phối màu của bài Mầm xanh tương lai với ạ.
                      createdAt: "2026-07-24T09:12:00+07:00"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/DemoIdentityRequired"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "503":
          $ref: "#/components/responses/RequestTimeout"
  /api/v1/assistant/conversations:
    get:
      tags:
        - Trợ lý
      summary: Lấy lịch sử chat với Artly
      description: |
        Trả về các hội thoại thuộc tài khoản hiện tại, mới hoạt động nhất trước.
        Hội thoại chỉ được tạo sau khi trợ lý trả lời thành công
        cho tin nhắn đầu tiên.
      operationId: listAssistantConversations
      security:
        - DemoUserId: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/AssistantConversationPageSize"
      responses:
        "200":
          description: Danh sách hội thoại và thông tin phân trang.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantConversationListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/DemoIdentityRequired"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /api/v1/assistant/conversations/{id}:
    parameters:
      - $ref: "#/components/parameters/AssistantConversationId"
    get:
      tags:
        - Trợ lý
      summary: Mở lại một lịch sử chat
      description: |
        Trả về metadata và toàn bộ tin nhắn theo đúng thứ tự. Nếu ID không tồn
        tại hoặc thuộc tài khoản khác, API trả 404 để không làm lộ dữ liệu.
      operationId: getAssistantConversation
      security:
        - DemoUserId: []
      responses:
        "200":
          description: Hội thoại và các tin nhắn đã lưu.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssistantConversationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/DemoIdentityRequired"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token của phiên Supabase Auth hiện tại.

    DemoUserId:
      type: apiKey
      in: header
      name: X-User-ID
      description: |
        UUID của một tài khoản mẫu trong bảng `users`, gửi ở dạng chuỗi chuẩn có
        dấu gạch ngang. Đây là cơ chế chọn danh tính cho MVP, không phải API key
        hoặc auth production.
      x-example: "00000000-0000-4000-8000-000000000001"
  parameters:
    Page:
      name: page
      in: query
      required: false
      description: Số trang, bắt đầu từ 1.
      schema:
        type: integer
        format: int32
        minimum: 1
        maximum: 100000
        default: 1
      example: 1
...
  schemas:
...
    UserResponse:
      type: object
      additionalProperties: false
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/User"
...